Move UniqueFields and ValueTransformer to attributes package
