+++ warehouse/query-core/src/main/java/datawave/query/jexl/visitors/QueryOptionsFromQueryVisitor.java
@@ -4,8 +4,8 @@ import com.google.common.base.Joiner;
 import com.google.common.collect.ImmutableSet;
 import datawave.query.QueryParameters;
 import datawave.query.jexl.functions.QueryFunctions;
-import datawave.query.transformer.UniqueFields;
-import datawave.query.transformer.ValueTransformer;
+import datawave.query.attributes.UniqueFields;
+import datawave.query.attributes.ValueTransformer;
 import org.apache.commons.jexl2.parser.ASTFunctionNode;
 import org.apache.commons.jexl2.parser.ASTIdentifier;
 import org.apache.commons.jexl2.parser.ASTReference;
