+++ warehouse/query-core/src/test/java/datawave/query/attributes/UniqueFieldsTest.java
@@ -1,4 +1,4 @@
-package datawave.query.transformer;
+package datawave.query.attributes;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
