+++ warehouse/query-core/src/test/java/datawave/query/config/ShardQueryConfigurationTest.java
@@ -18,8 +18,8 @@ import datawave.query.DocumentSerialization;
 import datawave.query.function.DocumentPermutation;
 import datawave.query.function.DocumentProjection;
 import datawave.query.model.QueryModel;
-import datawave.query.transformer.UniqueFields;
-import datawave.query.transformer.ValueTransformer;
+import datawave.query.attributes.UniqueFields;
+import datawave.query.attributes.ValueTransformer;
 import datawave.util.TableName;
 import datawave.webservice.query.QueryImpl;
 import org.junit.Assert;
